requirements.txt
psycopg2-binary
selenium
undetected-chromedriver
schedule
streamlit
pandas
requests
yfinance
pyvirtualdisplay